# docs/openapi_spec.json
{
...
  "info": {
    "description": "Comprehensive API for managing forms, responses, and AI analysis.",
    "title": "Forms Backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/form/api/v1/dashboard-settings/widgets/{widget_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "widget_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Remove a widget from the user's dashboard.\n\nArgs:\n    widget_id: ID of the widget to remove\n\nReturns:\n    200: Success message\n    404: Widget not found\n    401: Unauthorized",
        "tags": [
          "Dashboard_Settings"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "widget_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Update a widget's configuration.\n\nArgs:\n    widget_id: ID of the widget to update\n\nRequest Body:\n    {\n        \"position\": {\"x\": 0, \"y\": 4},  // Optional new position\n        \"size\": {\"w\": 2, \"h\": 2},      // Optional new size\n        \"config\": {...},               // Optional config updates\n        \"is_visible\": True             // Optional visibility\n    }\n\nReturns:\n    200: Updated widget object\n    404: Widget not found\n    401: Unauthorized",
        "tags": [
          "Dashboard_Settings"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}